moving Registry and PluginRegistry to components/registry.py
