+++ src/datumaro/__init__.py
@@ -37,7 +37,7 @@ from .components.dataset import Dataset, DatasetSubset, IDataset, eager_mode
 from .components.dataset_base import CategoriesInfo, DatasetBase, DatasetItem, SubsetBase
 from .components.dataset_item_storage import ItemStatus
 from .components.dataset_storage import DatasetPatch
-from .components.environment import Environment, PluginRegistry
+from .components.environment import Environment
 from .components.exporter import Exporter, ExportErrorPolicy, FailingExportErrorPolicy
 from .components.hl_ops import (  # pylint: disable=redefined-builtin
     export,
@@ -52,6 +52,7 @@ from .components.launcher import Launcher
 from .components.media import Image, MediaElement, PointCloud, Video, VideoFrame
 from .components.media_manager import MediaManager
 from .components.progress_reporting import NullProgressReporter, ProgressReporter
+from .components.registry import PluginRegistry
 from .components.transformer import ItemTransform, ModelTransform, Transform
 from .components.validator import Validator
 from .util.definitions import DEFAULT_SUBSET_NAME
